+++ src/ContentType/IgnoreContentTypeDefinition.cs
@@ -19,12 +19,52 @@ namespace IgnoreFiles
 
         [Export(typeof(FileExtensionToContentTypeDefinition))]
         [ContentType(IgnoreContentType)]
+        [FileExtension(".tfignore")]
+        public FileExtensionToContentTypeDefinition TfFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
         [FileExtension(".hgignore")]
         public FileExtensionToContentTypeDefinition HgFileExtension { get; set; }
 
         [Export(typeof(FileExtensionToContentTypeDefinition))]
         [ContentType(IgnoreContentType)]
+        [FileExtension(".npmignore")]
+        public FileExtensionToContentTypeDefinition NpmFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".dockerignore")]
+        public FileExtensionToContentTypeDefinition DockerFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".chefignore")]
+        public FileExtensionToContentTypeDefinition ChefFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".cvsignore")]
+        public FileExtensionToContentTypeDefinition CvsFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".bzrignore")]
+        public FileExtensionToContentTypeDefinition BazaarFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".jshintignore")]
+        public FileExtensionToContentTypeDefinition JshintFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
         [FileExtension(".eslintegnore")]
         public FileExtensionToContentTypeDefinition EslintFileExtension { get; set; }
+
+        [Export(typeof(FileExtensionToContentTypeDefinition))]
+        [ContentType(IgnoreContentType)]
+        [FileExtension(".cfignore")]
+        public FileExtensionToContentTypeDefinition CloudFoundryFileExtension { get; set; }
     }
 }