+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [x] Colorizer
 - [x] File icons for Solution Explorer
+- [x] Support all known .ignore files
 
 Features that have a checkmark are complete and available for
 download in the